% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/kidstats.R
\docType{package}
\name{kidstats}
\alias{kidstats}
\alias{kidstats-package}
\title{kidstats: Shiny application to make sub-adult age and sex estimations.}
\usage{
kidstats()
}
\description{
\code{kidstats} Shiny application to make sub-adult age and sex estimations
based on long-bone measurements using MARS models.

Kidstats shiny app call
}
\details{
This package contains the following datasets:
\itemize{
  \item \code{salb_za}: Sub-adult long bone length measurements
   (South Africa)
}

and the following functions:
\itemize{
  \item \code{kidstats}: Runs the kidstats shiny app.
}
}
\examples{
kidstats()
}

