% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salb_za}
\alias{salb_za}
\title{Sub-adult long bone length (South Africa)}
\format{A data frame with 1380 rows and 22 variables:}
\usage{
salb_za
}
\description{
Dissertation data collected by Kyra Stull PhD. Eighteen post-cranial
measurements were taken from Lodox Statscan images of 1380 South African
children aged birth through twelve.
}
\section{Variables}{

\itemize{
  \item \code{id}: Unique case ID
  \item \code{sex}: M or F
  \item \code{ancestry}: Ancestry
  \item \code{age_y}: Age at scan date (years)
  \item \code{h_mxl}: Humerus maximum length (mm)
  \item \code{h_pb}: Humerus proximal breadth (mm)
  \item \code{h_db}: Humerus distal breadth (mm)
  \item \code{h_msb}: Humerus mid-shaft breadth (mm)
  \item \code{u_mxl}: Ulna maximum length (mm)
  \item \code{u_msb}: Ulna mid-shaft breadth (mm)
  \item \code{r_mxl}: Radius maximum length (mm)
  \item \code{r_pb}: Radius proximal breadth (mm)
  \item \code{r_db}: Radius distal breadth (mm)
  \item \code{r_msb}: Radius mid-shaft breadth (mm)
  \item \code{f_mxl}: Femur maximum length (mm)
  \item \code{f_db}: Femur distal breadth (mm)
  \item \code{f_msb}: Femur mid-shaft breadth (mm)
  \item \code{t_mxl}: Tibia maximum length (mm)
  \item \code{t_pb}: Tibia proximal breadth (mm)
  \item \code{t_db}: Tibia distal breadth (mm)
  \item \code{t_msb}: Tibia mid-shaft breadth (mm)
  \item \code{fb_mxl}: Fibula maximum length (mm)
}
}
\keyword{datasets}

