% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{salb_za}
\alias{salb_za}
\title{Sub-adult long bone length (South Africa)}
\format{A data frame with 1380 rows and 22 variables:}
\usage{
salb_za
}
\description{
Dissertation data collected by Kyra Stull PhD. Eighteen post-cranial
measurements were taken from Lodox Statscan images of 1380 South African
children aged birth through twelve.
}
\section{Variables}{

\itemize{
  \item \code{ID}: Unique case ID
  \item \code{SEX}: M or F
  \item \code{ANCESTRY}: Ancestry
  \item \code{AGE}: Age at scan date (years)
  \item \code{HDL}: Humerus diaphyseal length (mm)
  \item \code{HPB}: Humerus proximal breadth (mm)
  \item \code{HDB}: Humerus distal breadth (mm)
  \item \code{HMSB}: Humerus mid-shaft breadth (mm)
  \item \code{UDL}: Ulna diaphyseal length (mm)
  \item \code{UMSB}: Ulna mid-shaft breadth (mm)
  \item \code{RDL}: Radius diaphyseal length (mm)
  \item \code{RPB}: Radius proximal breadth (mm)
  \item \code{RDB}: Radius distal breadth (mm)
  \item \code{RMSB}: Radius mid-shaft breadth (mm)
  \item \code{FDL}: Femur maximum length (mm)
  \item \code{FDB}: Femur distal breadth (mm)
  \item \code{FMSB}: Femur mid-shaft breadth (mm)
  \item \code{TDL}: Tibia diaphyseal length (mm)
  \item \code{TPB}: Tibia proximal breadth (mm)
  \item \code{TDB}: Tibia distal breadth (mm)
  \item \code{TMSB}: Tibia mid-shaft breadth (mm)
  \item \code{FBDL}: Fibula diaphyseal length (mm)
}
}
\keyword{datasets}

